Renamed Imtls -> DictArray
