+++ debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Renamed Imtls -> DictArray
   * Removed the .id attribute from the Site class
 
   [Graeme Weatherill]
